backend/pyproject.toml
[tool.poetry]
name = "dofus-market"
version = "0.1.0"
description = "The goal of this project is to streamline theory crafting for dofus touch"
authors = ["Ludovic CHAPELET <ludovic.chapelet@outlook.com>"]
license = "APACHE 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
django-filter = "^24.2"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
aiohttp = "^3.9.5"
psycopg2-binary = "^2.9.9"
django-dotenv = "^1.4.2"
django-socio-grpc = "^0.22.8"
pydantic = "^2.8.2"
django-cors-headers = "^4.4.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.5.2"

[tool.poetry.group.tests.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
pytest-django = "^4.8.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
